requirements.txt
Flask==3.0.3
pymongo==3.12.1
python-telegram-bot==21.2
aiohttp==3.9.5
pydantic~=2.7.1
python-dateutil~=2.9.0.post0